% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kendrick.R
\name{add_kendrick}
\alias{add_kendrick}
\title{Calculate kendrick mass}
\usage{
add_kendrick(ftmsObj, base_compounds = "CH2")
}
\arguments{
\item{data:fticrdata, }{base compound: repeating unit}
}
\value{
add_kendrick mass and KMD
}
\description{
using use data
}
