% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_anova.R
\name{anova_stat}
\alias{anova_stat}
\title{Perform anova repeatly}
\usage{
anova_stat(data, grpcol = 1, varcol = 3)
}
\arguments{
\item{data:input}{data, grpcol:column number of group, varcol: column of number of variable}
}
\value{
anova F.value and P.value each variable
}
\description{
using use data
}
