% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.N_ratio.R
\name{S.N_ratio}
\alias{S.N_ratio}
\title{Ignore raman signal using threshold}
\usage{
S.N_ratio(data, Threshold = 0.001)
}
\arguments{
\item{data, }{spectrumdata and set threshold}
}
\value{
check ramna intensity whether signal or not and noise clearing
}
\description{
using melt data
}
