% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stat.R
\name{two_stat}
\alias{two_stat}
\title{Perform t-test repeatly}
\usage{
two_stat(data, grpcol = 1, varcol = 3, var.opt = T)
}
\arguments{
\item{data:input}{data, grpcol:column number of group, varcol: column of number of variable, var.opt=Var.equal in t.test}
}
\value{
compare two group with p.val according to variable
}
\description{
using use data
}
